+++ apps/user/src/apis/token/refresh/index.ts
@@ -2,7 +2,7 @@ import { LOGIN_PAGE_ROUTE } from '@/constants/routes';
 import { ACCESS_KEY, REFRESH_KEY } from '@/constants/token';
 import { useRouter } from 'next/navigation';
 import { maru } from '../../instance';
-import { Storage } from '../../storage/storage';
+import { Storage } from '../../storage';
 
 export const refreshToken = async () => {
     const router = useRouter();
